pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "planetsca"
dynamic = ["version"]
readme = "README.md"
description = "Mapping snow covered areas (SCA) from high-resolution PlanetScope images using Random Forest model"
dependencies = [
    "dvc>=2.10.0",
    "hydra-core>=1.1.1",
    "numpy",
    "huggingface-hub>=0.23.4",
    "pdoc3>=0.10.0",
    "pandas",
    "scikit-learn",
    "rasterio",
    "matplotlib",
    "joblib>=1.3.2",
    "requests",
    "shapely",
    "geopandas",
    "fiona",
    "skl2onnx",
    "onnxruntime",

]
requires-python = ">=3.8"
authors = [
    {name = "Ian Chiu", email = "ianchiu333@gmail.com"},
    {name = "Nicoleta Cristea", email = "cristn@uw.edu"},
    {name = "Steven Pestana", email = "spestana@uw.edu"},
    {name = "Kehan Yang", email = "kyang33@uw.edu"},
]
license = {file = "LICENSE"}


[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "pre-commit>=2.17.0",
    "nox",
]
doc = [
    "pandoc",
    "sphinx",
    "furo",
    "myst-parser",
    "nbsphinx",
    "ipykernel",
    "ghp-import",
]

[project.urls]
Homepage = "https://test.pypi.org/project/PyPlanetSCA/#description"
Repository = "https://github.com/DSHydro/PyPlanetSCA-Python-Library"

[tool.hatch.build.targets.wheel]
packages = ["src/planetsca"]

[tool.hatch.build.targets.sdist]
exclude = [
  "/tests",
  "/code_archive",
  "/demo",
  "/additional_assets"
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/planetsca/version.py"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]
